+++ src/030-error/display-to-json.ts
