+++ src/bindings/typescript/writeStruct.ts
@@ -1,7 +1,7 @@
 import { ABIType, ABITypeRef } from "@ton/core";
 import { serializers } from "./serializers";
 import { AllocationCell, AllocationOperation } from "../../storage/operation";
-import { throwInternalCompilerError } from "../../error/errors";
+import { throwInternalCompilerError } from "../../030-error/errors";
 import { Writer } from "../../utils/Writer";
 
 export const maxTupleSize = 15;
