+++ src/grammar/prev/grammar.ts
@@ -1,6 +1,6 @@
 import { Node, IterationNode, NonterminalNode } from "ohm-js";
 import tactGrammar from "./grammar.ohm-bundle";
-import { throwInternalCompilerError } from "../../error/errors";
+import { throwInternalCompilerError } from "../../030-error/errors";
 import {
     AstAugmentedAssignOperation,
     AstConstantAttribute,
@@ -19,7 +19,7 @@ import {
     FactoryAst,
 } from "../ast";
 import { ItemOrigin, SrcInfo } from "../src-info";
-import { displayToString } from "../../error/display-to-string";
+import { displayToString } from "../../030-error/display-to-string";
 import { ParserErrors, parserErrorSchema } from "./parser-error";
 import { getSrcInfoFromOhm } from "./src-info";
