+++ src/interpreter/interpreter.ts
@@ -9,7 +9,7 @@ import {
     idTextErr,
     throwConstEvalError,
     throwInternalCompilerError,
-} from "../error/errors";
+} from "../030-error/errors";
 import {
     AstBinaryOperation,
     AstBoolean,
