+++ src/main.ts
@@ -16,4 +16,4 @@ export { createVirtualFileSystem } from "./020-vfs/createVirtualFileSystem";
 export * from "./browser";
 export * from "./verify";
 export * from "./010-pipeline/logger";
-export * from "./error/errors";
+export * from "./030-error/errors";
