+++ src/storage/allocator.ts
@@ -4,7 +4,7 @@ import {
     AllocationOperation,
     AllocationOperationType,
 } from "./operation";
-import { throwInternalCompilerError } from "../error/errors";
+import { throwInternalCompilerError } from "../030-error/errors";
 
 const ALLOCATOR_RESERVE_BIT = 1;
 const ALLOCATOR_RESERVE_REF = 1;
